// frontend/src/api/generated/serializers/Membership.ts
// Typelizer digest 986f37cf2c0c874c45f741c48bddb8fc
//
// DO NOT MODIFY: This file was automatically generated by Typelizer.
import type {User} from "./index"

type Membership = {
  id: string;
  role: "owner" | "admin" | "member";
  access: "full" | "viewer";
  insertedAt: string;
  user: User | null;
}

export type { Membership };